Updated resources for REST API

+++ resources/item.py
@@ -1,7 +1,7 @@
-import uuid
-from flask import request
+
 from flask.views import MethodView
 from flask_smorest import Blueprint, abort 
+from flask_jwt_extended import jwt_required, get_jwt
 from schemas import ItemSchema, ItemUpdateSchema
 from models import ItemModel
 from db import db
@@ -12,17 +12,24 @@ blp = Blueprint("items", __name__, description="Operations on items")
 @blp.route("/item/<int:item_id>")
 class Item(MethodView):
     
+    @jwt_required()
     @blp.response(200, ItemSchema)
     def get(self, item_id):
       item = ItemModel.query.get_or_404(item_id)
       return item
 
+    @jwt_required()
     def delete(self, item_id):
+        jwt = get_jwt()
+        if not jwt["is_admin"]:
+            abort(403, message="Admin privilege required")
+            
         item = ItemModel.query.get_or_404(item_id)
         db.session.delete(item)
         db.session.commit()
         return {"message": "Item deleted"}
 
+    @jwt_required()
     @blp.arguments(ItemUpdateSchema)
     @blp.response(200, ItemSchema)
     def put(self, item_data, item_id):
@@ -40,10 +47,12 @@ class Item(MethodView):
 @blp.route("/item")
 class ItemList(MethodView):
     
+    @jwt_required()
     @blp.response(200, ItemSchema(many=True))
     def get(self):
         return ItemModel.query.all()
     
+    @jwt_required()
     @blp.arguments(ItemSchema)
     @blp.response(201, ItemSchema)
     def post(self, item_data):
